refactor: useAnime composable split
