+++ app/components/ui/input/Input.vue
@@ -31,5 +31,5 @@ const modelValue = useVModel(props, 'modelValue', emits, {
         props.class
       )
     "
-  >
+  />
 </template>
