+++ app/utils/anime/cache.ts
@@ -1,0 +1,27 @@
+import type { Media } from '~/utils/types/anilist'
+import type { CacheEntry } from '~/utils/types/anime'
+
+export const generateCacheKey = (
+  searchQuery: string,
+  selectedSort: string,
+  selectedSeason: string
+): string => {
+  return `${searchQuery}-${selectedSort}-${selectedSeason}`
+}
+
+export const getCachedData = (
+  cache: Record<string, CacheEntry>,
+  cacheKey: string
+): CacheEntry | undefined => {
+  return cache[cacheKey]
+}
+
+export const setCacheData = (
+  cache: Record<string, CacheEntry>,
+  cacheKey: string,
+  data: Media[],
+  page: number,
+  hasMore: boolean
+): void => {
+  cache[cacheKey] = { data, page, hasMore }
+}
