+++ nuxt.config.ts
@@ -114,5 +114,5 @@ export default defineNuxtConfig({
   // TypeScript configuration
   typescript: {
     typeCheck: false // need to add vue-tsc dep for prod build when true
-  },
+  }
 })
